backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart